@@ -2,6 +2,10 @@
     "name": "coffeephp/extension",
     "description": "Auto-loads extensions that add functionality to PHP.",
     "license": "Apache-2.0",
+    "version": "1.1.0",
+    "type": "library",
+    "readme": "README.md",
+    "time": "2021-01-08",
     "authors": [
         {
             "name": "Danny Damsky",
@@ -15,7 +19,7 @@
         "extension"
     ],
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "ext-apcu": "*",
         "ext-mbstring": "*"
     },
@@ -41,22 +45,20 @@
     "config": {
         "sort-packages": true,
         "optimize-autoloader": true,
-        "apcu-autoloader": true
+        "apcu-autoloader": true,
+        "platform": {
+            "php": "8.0"
+        },
+        "platform-check": false
     },
     "scripts": {
         "coffeephp:exec:phpcbf": "coffeephp_exec_phpcbf",
         "coffeephp:exec:phpcs": "coffeephp_exec_phpcs",
         "coffeephp:exec:phpmd": "coffeephp_exec_phpmd",
         "coffeephp:exec:phpstan-analyse": "coffeephp_exec_phpstan_analyse",
-        "coffeephp:exec:phpunit": "coffeephp_exec_phpunit",
+        "coffeephp:exec:phpunit": "coffeephp_exec_phpunit --testsuite unit,integration",
         "coffeephp:exec:psalm": "coffeephp_exec_psalm",
         "coffeephp:exec:psalm-alter": "coffeephp_exec_psalm_alter",
-        "coffeephp:exec:publish": "coffeephp_exec_publish",
-        "coffeephp:exec:analyse": [
-            "@coffeephp:exec:phpunit",
-            "@coffeephp:exec:psalm",
-            "@coffeephp:exec:phpcs",
-            "@coffeephp:exec:phpstan-analyse"
-        ]
+        "coffeephp:exec:publish": "coffeephp_exec_publish"
     }
 }
